requirements.txt
Pillow==9.4.0
pygame==2.2.0
PyTMX==3.31
